% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\docType{data}
\name{channelInfo}
\alias{channelInfo}
\title{Channel information}
\format{
a data frame with 18 rows and 7 variables
}
\usage{
data(channelInfo)
}
\description{
Spectral response information on the 18 channels of the AS7265x. The 
variables are as follows:
\itemize{ 
\item channel name of channel 
\item channelIndex order (index) of channel as returned by the device
\item wavelength primary wavelength in nm
\item color hex value of color 
\item red red value for RGB approximation 
\item green green value for RGB approximation 
\item blue blue value for RGB approximation 
}
}
