% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\docType{data}
\name{colorMatch}
\alias{colorMatch}
\title{CIE color matching data
 
Data for conversion of visible light wavelengths to 
relative stimulation of long, medium and short human cone 
cells. The variables are as follows: 
\itemize{ 
\item wavelength 
\item xbar 
\item ybar 
\item zbar
} 
These data were obtained from the reference below.}
\format{
a data frame with 470 rows and 4 variables
}
\usage{
data(colorMatch)
}
\description{
CIE color matching data
 
Data for conversion of visible light wavelengths to 
relative stimulation of long, medium and short human cone 
cells. The variables are as follows: 
\itemize{ 
\item wavelength 
\item xbar 
\item ybar 
\item zbar
} 
These data were obtained from the reference below.
}
\references{
\url{https://cie.co.at/datatable/cie-1931-colour-matching-functions-2-degree-observer}
}
