% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{confirmCom}
\alias{confirmCom}
\title{Confirm command}
\usage{
confirmCom(timeout = 1)
}
\arguments{
\item{timeout}{time to wait for a response in seconds}
}
\description{
Obtain response from an AT command.
}
\details{
An AT command that doesn't return data should return either 'OK' or 'ERROR'.
This command waits for one of these two responses.
}
\author{
Peter Hoff
}
