% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{getEEM}
\alias{getEEM}
\title{Get EEM}
\usage{
getEEM(plot = FALSE)
}
\arguments{
\item{plot}{(logical) plot the spectrum}
}
\description{
Get excitation-emmission matrix (EEM).
}
\details{
Eighteen spectral readings are obtained separately under each of the three 
illumination bulbs (ultraviolet, visible light, infraread), 
and the results are returned as a 3 by 18 matrix.
}
\author{
Peter Hoff
}
