% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{getSpec}
\alias{getSpec}
\title{Get spectrum}
\usage{
getSpec(leds = c("UV", "VI", "IR"))
}
\arguments{
\item{leds}{which LEDs to use for illumination}
}
\value{
an 18-dimensional vector of light intensity readings
}
\description{
Get spectral data from the device.
}
\details{
The device returns 18-channel light intensity readings 
corresponding to wavelengths ranging from 
ultraviolet to infrared. Up to three different 
LEDs can be used to illuminate the sample.
}
\author{
Peter Hoff
}
