% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{getSpecMat}
\alias{getSpecMat}
\title{Get spectral matrix}
\usage{
getSpecMat(plot = FALSE)
}
\arguments{
\item{plot}{(logical) plot the spectrum}
}
\description{
Get spectrum under multiple illuminations.
}
\details{
Eighteen spectral readings are obtained separately under each of the three 
illumination bulbs (ultraviolet, visible light, infrared), 
and the results are returned as a 3 by 18 matrix.
}
\author{
Peter Hoff
}
