% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{getSpecMats}
\alias{getSpecMats}
\title{Get spectral matrices}
\usage{
getSpecMats(n = 5)
}
\arguments{
\item{n}{(integer) number of replicate scans per sample}
}
\description{
Get spectral matricies for multiple samples.
}
\details{
Interactively collects SMs for a series of samples, 
allowing for naming of each sample.
}
\author{
Peter Hoff
}
