% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{plotSpec}
\alias{plotSpec}
\title{Spectrum plot}
\usage{
plotSpec(s)
}
\arguments{
\item{s}{numeric vector giving the spectrum, with names equal to wavelengthts}
}
\description{
Plots a spectrum as a bar plot
}
\author{
Peter Hoff
}
