% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{spec2rgb}
\alias{spec2rgb}
\title{Spectrum to RGB}
\usage{
spec2rgb(s)
}
\arguments{
\item{s}{a vector giving a spectrum, with names giving the wavelengths}
}
\description{
Convert spectrum to XYZ values and then to RGB
}
\details{
A spectrum is detected by the human eye by three sensors, 
stimulation of which is then converted by the brain  
into color perception. 
This function provides RGB values for the perceived 
color.
}
\references{
\url{https://www.fourmilab.ch/documents/specrend}
}
\author{
Peter Hoff
}
