% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{specClose}
\alias{specClose}
\title{Close spectrometer connection}
\usage{
specClose()
}
\description{
Close serial connection to the spectrometer.
}
\details{
An existing connection to the serial device is closed.
}
\author{
Peter Hoff
}
