% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{specOpen}
\alias{specOpen}
\title{Open spectrometer connection}
\usage{
specOpen()
}
\description{
Open serial connection to the spectrometer.
}
\details{
A serial device is identified, and a connection is made to it.
}
\author{
Peter Hoff
}
